backend/requirements.txt
flask
flask-cors
flask-mail
mysql-connector-python
flask-jwt-extended
fpdf
pillow
numpy
tflite-runtime
